chore: Move AllowAllPermissionPolicy to it's own backend module package
