+++ packages/backend-next/package.json
@@ -41,6 +41,7 @@
     "@backstage/plugin-lighthouse-backend": "workspace:^",
     "@backstage/plugin-linguist-backend": "workspace:^",
     "@backstage/plugin-permission-backend": "workspace:^",
+    "@backstage/plugin-permission-backend-module-policy-allow-all": "workspace:^",
     "@backstage/plugin-permission-common": "workspace:^",
     "@backstage/plugin-permission-node": "workspace:^",
     "@backstage/plugin-proxy-backend": "workspace:^",
