+++ plugins/permission-backend-module-policy-allow-all/README.md
@@ -1,0 +1,5 @@
+# @backstage/plugin-permission-backend-module-policy-allow-all
+
+The allow all policy backend module for the permission plugin.
+
+_This plugin was created through the Backstage CLI_
