+++ plugins/permission-backend-module-policy-allow-all/src/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,4 +14,10 @@
  * limitations under the License.
  */
 
-export { permissionPlugin } from './plugin';
+/**
+ * The allow all policy backend module for the permission plugin.
+ *
+ * @packageDocumentation
+ */
+
+export { permissionModuleAllowAllPolicy } from './module';
