+++ plugins/permission-backend-module-policy-allow-all/src/policy.ts
@@ -1,5 +1,15 @@
+import { BackstageIdentityResponse } from '@backstage/plugin-auth-node';
+import {
+  AuthorizeResult,
+  PolicyDecision,
+} from '@backstage/plugin-permission-common';
+import {
+  PermissionPolicy,
+  PolicyQuery,
+} from '@backstage/plugin-permission-node';
+
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,5 +23,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export { permissionPlugin } from './plugin';
+export class AllowAllPermissionPolicy implements PermissionPolicy {
+  async handle(
+    _request: PolicyQuery,
+    _user?: BackstageIdentityResponse,
+  ): Promise<PolicyDecision> {
+    return {
+      result: AuthorizeResult.ALLOW,
+    };
+  }
+}
